+++ src/containers/Effect/Effect.tsx
@@ -9,12 +9,12 @@ import {
 
 
 export const EffectContainer = props => {
-    const dispatch = useContext(Context);
+    const context = useContext(Context);
 
     return (
         <Effect
             {...props}
-            onParamChange={setSendParamValue(dispatch, props.id)}
+            onParamChange={setSendParamValue(context, props.id)}
         />
     );
 };
