+++ src/containers/index.ts
@@ -2,3 +2,4 @@ export * from './Context';
 export * from './Desk';
 export * from './Effect';
 export * from './Track';
+export * from './Mixdesk';
